flashspotter/requirements.txt
eth-brownie==1.20.1
requests
